+++ src/Frontend/Dfe.Complete/Pages/Projects/Decision/RecordDaoRevocation/DaoRevocationProjectLayoutModel.cs
@@ -1,0 +1,96 @@
+﻿using Dfe.Complete.Application.DaoRevoked.Commands;
+using Dfe.Complete.Constants;
+using Dfe.Complete.Domain.Enums;
+using Dfe.Complete.Domain.ValueObjects;
+using Dfe.Complete.Pages.Projects.ProjectView;
+using Dfe.Complete.Services;
+using Dfe.Complete.Utils;
+using GovUK.Dfe.CoreLibs.Caching.Helpers;
+using GovUK.Dfe.CoreLibs.Caching.Interfaces;
+using MediatR;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.ModelBinding;
+using System.Diagnostics.CodeAnalysis;
+
+namespace Dfe.Complete.Pages.Projects.Decision.RecordDaoRevocation
+{
+    [ExcludeFromCodeCoverage]
+    public abstract class DaoRevocationProjectLayoutModel(ISender sender, ILogger logger, ICacheService<IMemoryCacheType> cacheService) : ProjectLayoutModel(sender, logger, RecordDaoRevocationNavigation)
+    {
+        protected string CacheKey => $"DaoRevocation_{CacheKeyHelper.GenerateHashedCacheKey(ProjectId)}";
+
+        protected IActionResult ReturnPage(RecordDaoRevocationDecisionCommand decision)
+            => decision.ReasonNotes?.Count == 0 ? RedirectToDaoRevocationPage() : Page();
+
+        protected async Task<RecordDaoRevocationDecisionCommand> GetCachedDecisionAsync()
+        {
+            return await cacheService.GetOrAddAsync(CacheKey, () =>
+            {
+                var command = new RecordDaoRevocationDecisionCommand(new ProjectId(new Guid(ProjectId)));
+                return Task.FromResult(command);
+            }, string.Empty);
+        }
+
+        protected async Task UpdateCacheAsync(RecordDaoRevocationDecisionCommand decison)
+        { 
+            cacheService.Remove(CacheKey); 
+            await cacheService.GetOrAddAsync(CacheKey, () =>
+            { 
+                return Task.FromResult(decison);
+            }, string.Empty);
+        }
+        public RedirectResult RedirectToDaoRevocationPage()
+            => RedirectToDaoRevocationRoute(RouteConstants.ProjectDaoRevocation);
+
+        public string GetDaoRevocationCheckPage()
+            => FormatRouteWithProjectId(RouteConstants.ProjectDaoRevocationCheck);
+
+        protected RedirectResult RedirectToDaoRevocationRoute(string route)
+            => Redirect(FormatRouteWithProjectId(route));
+
+        protected static void PoplateOptions(List<DaoRevokedReason> reasons)
+        {
+            reasons.Add(DaoRevokedReason.SchoolRatedGoodOrOutstanding);
+            reasons.Add(DaoRevokedReason.SafeguardingConcernsAddressed);
+            reasons.Add(DaoRevokedReason.SchoolClosedOrClosing);
+            reasons.Add(DaoRevokedReason.ChangeToGovernmentPolicy);
+        }
+        protected static void ValidateReasons(IFormCollection formValues, List<DaoRevokedReason> reasons, Dictionary<DaoRevokedReason, string> reasonNotes, ErrorService errorService, ModelStateDictionary modelState)
+        {
+            var errors =new Dictionary<string, string>();
+            foreach (var reason in reasons)
+            {
+                var reasonKey = $"dao_revoked_reasons[{reason.ToDescription()}]";
+                var noteKey = $"dao_revoked_reasons[{reason.ToDescription()}_note]";
+
+                var isChecked = formValues.TryGetValue(reasonKey, out var selected) && selected == "1";
+                if (!isChecked)
+                    continue;
+
+                var hasNote = formValues.TryGetValue(noteKey, out var note);
+                var isNoteValid = hasNote && !string.IsNullOrWhiteSpace(note);
+
+                if (isNoteValid)
+                {
+                    reasonNotes[reason] = note!;
+                }
+                else
+                {
+                    errors.Add($"{reason.ToDescription()}_note", ValidationConstants.MustProvideDetails);
+                }
+            }
+
+            if (reasonNotes.Count == 0 && errors.Count == 0)
+            {
+                errors.Add("select-dao-revoked-reason", ValidationConstants.ChooseAtLeastOneReason);
+            } 
+
+            foreach (var error in errors.Distinct())
+            {
+                modelState.AddModelError(error.Key, error.Value);
+            }
+
+            errorService.AddErrors(modelState);
+        }
+    }
+}